Presenter code was in MapsView and View code was in MapsPresenter. All dependent libraries are updated and Android Api level updated to 33 to 34
