+++ app/src/main/java/me/amitshekhar/ridesharing/ui/maps/MapsView.kt
@@ -1,27 +1,127 @@
 package me.amitshekhar.ridesharing.ui.maps
 
+import android.util.Log
 import com.google.android.gms.maps.model.LatLng
+import me.amitshekhar.ridesharing.data.network.NetworkService
+import me.amitshekhar.ridesharing.simulator.WebSocket
+import me.amitshekhar.ridesharing.simulator.WebSocketListener
+import me.amitshekhar.ridesharing.utils.Constants
+import org.json.JSONObject
 
-interface MapsView {
+class MapsView(private val networkService: NetworkService) : WebSocketListener {
 
-    fun showNearbyCabs(latLngList: List<LatLng>)
+    companion object {
+        private const val TAG = "MapsPresenter"
+    }
 
-    fun informCabBooked()
+    private var view: MapsPresenter? = null
+    private lateinit var webSocket: WebSocket
 
-    fun showPath(latLngList: List<LatLng>)
+    fun onAttach(view: MapsPresenter?) {
+        this.view = view
+        webSocket = networkService.createWebSocket(this)
+        webSocket.connect()
+    }
 
-    fun updateCabLocation(latLng: LatLng)
+    fun onDetach() {
+        webSocket.disconnect()
+        view = null
+    }
 
-    fun informCabIsArriving()
+    fun requestNearbyCabs(latLng: LatLng) {
+        val jsonObject = JSONObject()
+        jsonObject.put(Constants.TYPE, Constants.NEAR_BY_CABS)
+        jsonObject.put(Constants.LAT, latLng.latitude)
+        jsonObject.put(Constants.LNG, latLng.longitude)
+        webSocket.sendMessage(jsonObject.toString())
+    }
 
-    fun informCabArrived()
+    fun requestCab(pickUpLatLng: LatLng, dropLatLng: LatLng) {
+        val jsonObject = JSONObject()
+        jsonObject.put("type", "requestCab")
+        jsonObject.put("pickUpLat", pickUpLatLng.latitude)
+        jsonObject.put("pickUpLng", pickUpLatLng.longitude)
+        jsonObject.put("dropLat", dropLatLng.latitude)
+        jsonObject.put("dropLng", dropLatLng.longitude)
+        webSocket.sendMessage(jsonObject.toString())
+    }
 
-    fun informTripStart()
+    private fun handleOnMessageNearbyCabs(jsonObject: JSONObject) {
+        val nearbyCabLocations = arrayListOf<LatLng>()
+        val jsonArray = jsonObject.getJSONArray(Constants.LOCATIONS)
+        for (i in 0 until jsonArray.length()) {
+            val lat = (jsonArray.get(i) as JSONObject).getDouble(Constants.LAT)
+            val lng = (jsonArray.get(i) as JSONObject).getDouble(Constants.LNG)
+            val latLng = LatLng(lat, lng)
+            nearbyCabLocations.add(latLng)
+        }
+        view?.showNearbyCabs(nearbyCabLocations)
+    }
 
-    fun informTripEnd()
+    override fun onConnect() {
+        Log.d(TAG, "onConnect")
+    }
 
-    fun showRoutesNotAvailableError()
+    override fun onMessage(data: String) {
+        Log.d(TAG, "onMessage data : $data")
+        val jsonObject = JSONObject(data)
+        when (jsonObject.getString(Constants.TYPE)) {
+            Constants.NEAR_BY_CABS -> {
+                handleOnMessageNearbyCabs(jsonObject)
+            }
+            Constants.CAB_BOOKED -> {
+                view?.informCabBooked()
+            }
+            Constants.PICKUP_PATH, Constants.TRIP_PATH -> {
+                val jsonArray = jsonObject.getJSONArray("path")
+                val pickUpPath = arrayListOf<LatLng>()
+                for (i in 0 until jsonArray.length()) {
+                    val lat = (jsonArray.get(i) as JSONObject).getDouble("lat")
+                    val lng = (jsonArray.get(i) as JSONObject).getDouble("lng")
+                    val latLng = LatLng(lat, lng)
+                    pickUpPath.add(latLng)
+                }
+                view?.showPath(pickUpPath)
+            }
+            Constants.LOCATION -> {
+                val latCurrent = jsonObject.getDouble("lat")
+                val lngCurrent = jsonObject.getDouble("lng")
+                view?.updateCabLocation(LatLng(latCurrent, lngCurrent))
+            }
+            Constants.CAB_IS_ARRIVING -> {
+                view?.informCabIsArriving()
+            }
+            Constants.CAB_ARRIVED -> {
+                view?.informCabArrived()
+            }
+            Constants.TRIP_START -> {
+                view?.informTripStart()
+            }
+            Constants.TRIP_END -> {
+                view?.informTripEnd()
+            }
+        }
+    }
 
-    fun showDirectionApiFailedError(error: String)
+    override fun onDisconnect() {
+        Log.d(TAG, "onDisconnect")
+    }
+
+    override fun onError(error: String) {
+        Log.d(TAG, "onError : $error")
+        val jsonObject = JSONObject(error)
+        when (jsonObject.getString(Constants.TYPE)) {
+            Constants.ROUTES_NOT_AVAILABLE -> {
+                view?.showRoutesNotAvailableError()
+            }
+            Constants.DIRECTION_API_FAILED -> {
+                view?.showDirectionApiFailedError(
+                    "Direction API Failed : " + jsonObject.getString(
+                        Constants.ERROR
+                    )
+                )
+            }
+        }
+    }
 
 }
